+++ components/SearchBox/index.tsx
@@ -5,7 +5,7 @@ import { useTranslations } from 'next-intl';
 import { useState, type FC } from 'react';
 
 import { SearchBox } from '@/components/SearchBox/components/SearchBox';
-import { useKeyboardCommands } from '@/hooks/react-client/useKeyboardCommands';
+import { useKeyboardCommands } from '@/hooks/react-client';
 
 import styles from './index.module.css';
