+++ components/SearchPage/index.tsx
@@ -7,7 +7,7 @@ import { useEffect, useState, type FC } from 'react';
 
 import type { SearchDoc } from '@/components/SearchBox/components/SearchBox';
 import { pathToBreadcrumbs } from '@/components/SearchBox/utils';
-import { useBottomScrollListener } from '@/hooks/react-client/useBottomScrollListener';
+import { useBottomScrollListener } from '@/hooks/react-client';
 import { DEFAULT_ORAMA_QUERY_PARAMS } from '@/next.constants.mjs';
 import { orama, highlighter } from '@/next.orama.mjs';
