+++ hooks/react-client/index.ts
@@ -3,3 +3,6 @@ export { default as useDetectOS } from './useDetectOS';
 export { default as useMediaQuery } from './useMediaQuery';
 export { default as useNotification } from './useNotification';
 export { default as useClientContext } from './useClientContext';
+export { default as useKeyboardCommands } from './useKeyboardCommands';
+export { default as useClickOutside } from './useClickOutside';
+export { default as useBottomScrollListener } from './useBottomScrollListener';
